@@ -1,43 +1,36 @@
+/*
+ *    Copyright 2019 Maarten de Goede
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
 package eu.insertcode.portfolio
 
-import android.content.Context
-import android.net.ConnectivityManager
+import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import eu.insertcode.portfolio.comms.GetDataAsync
-import java.lang.ref.WeakReference
+import androidx.appcompat.app.AppCompatActivity
+
 
 /**
- * Created by MrTheGood.
- * Copyright © 2018 MrTheGood. All rights reserved.
+ * Created by maartendegoede on 21/02/2019.
+ * Copyright © 2019 insertCode.eu. All rights reserved.
  */
 class SplashActivity : AppCompatActivity() {
-    var paused = true
-        private set
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-    }
 
-    override fun onStart() {
-        super.onStart()
-        GetDataAsync(WeakReference(this)).execute(resources.getString(R.string.url_getData))
-    }
-
-    override fun onResume() {
-        super.onResume()
-        paused = false
-    }
-
-    override fun onPause() {
-        super.onPause()
-        paused = true
-    }
-
-    fun isNetworkConnected(): Boolean {
-        val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        val networkInfo = connectivityManager.activeNetworkInfo
-        return networkInfo != null && networkInfo.isConnected
+        startActivity(Intent(this, MainActivity::class.java))
+        finish()
     }
 }
